+++ CHANGELOG.md
@@ -7,8 +7,10 @@ This project adheres to [Semantic Versioning 2.0.0](https://semver.org/spec/v2.0
 ## [Unreleased]
 
 ### Added
+- Management and correct visualization of **overlapping events**
 
 ### Changed
+- in week view the overlapping tasks are grouped and layered, z-index ordered by start-time and duration (similar to Google calendar visualization)
 
 ### Deprecated